+++ CHANGELOG.md
@@ -45,6 +45,7 @@ If you're still stuck on 3.4 and find something in the latest `unpythonic` 0.14.
 - `Shim`: A shim holds a `box` or a `ThreadLocalBox`, and redirects attribute accesses to whatever object is currently in the box.
 - `islice` now supports negative start and stop. (**Caution**: no negative step; and it must consume the whole iterable to determine where it ends, if at all.)
 - `async_raise`: Inject KeyboardInterrupt into an arbitrary thread. (*CPython only*.)
+- `resolve_bindings`: Get the parameter bindings a given callable would establish if it was called with the given args and kwargs. This is mainly of interest for implementing memoizers, since this allows them to see (e.g.) `f(1)` and `f(a=1)` as the same thing for `def f(a): pass`.
 
 **Non-breaking changes**: