+++ CHANGELOG.md
@@ -65,7 +65,7 @@ If you're still stuck on 3.4 and find something in the latest `unpythonic` 0.14.
 - Use standard Python semantics for negative indices ([#6](https://github.com/Technologicat/unpythonic/issues/6)).
 - Escape continuation analysis in `unpythonic.syntax.util` now interprets also the literal name `throw` as invoking an escape continuation.
 - Fix pickling of `frozendict` ([#55](https://github.com/Technologicat/unpythonic/issues/55)).
-- Fix spurious cache misses in memoizers ([#26](https://github.com/Technologicat/unpythonic/issues/26)).
+- Fix spurious cache misses in memoizers ([#26](https://github.com/Technologicat/unpythonic/issues/26)). The bug affected `memoize`, `gmemoize`, `fix` and `fixtco`.
 
 ---